+++ create_env.py
@@ -16,12 +16,11 @@ class CreateEnvDatabase:
                 user=conf.DB_USER,
                 password=conf.DB_PASSWORD,
                 port=conf.DB_PORT,
-                ssl_context=conf.DB_SSL,
+                ssl=conf.DB_SSL,
                 database=conf.DB_NAME
             )
         except:
             print('Can not connect to database.')
-            raise
 
     def __enter__(self):
         return self